// block comments allowed on indented code with smarttabs
pull request for issue #547

This update allows for the following if the smarttabs option has been enabled:

	//[space]if(something) {
	//[space][tab]doSomething();
	//[space]}

... which is the result of block commenting out the following code in some text editors:

	if(something) {
	[tab]doSomething();
	}

The smarttabs regex is now doing a negative look-behind for a "//" prior to a [space][tab], and allowing it in that case.
Test cases have been updated to account for this addition as well.